requirements.txt
flask
python-dotenv
pyodbc
openai
pandas
qdrant-client
google-generativeai
pymongo
certifi